Adds ingress-nginx template (#74)
* Adds ingress-nginx template

This template adds a service that acts as a proxy for all incoming traffic to the cluster and redirects it to the corresponding services.

* Updating all template files before deployment template

* Updates deployment file

* Removes unused namespace

* Prepare templates for the webhook

* Extracts variables to template definition

* Adds variable expansion snippet

* Adds files to template in correct order

* Fixes manifest
- Typo in IMAGE PULL POLICY
- Double // in namespace/resource name
- Typo in job-patch folder
- Name of webhook validating configuration

* Uses Cluster traffic policy
We were having issues with Digital Ocean where the
load balancer was unhealthy and 1 of the 2 nodes
was ALWAYS marked as down. According to their
documentation one reason could be that external
health checks were not hitting. This option
changes this & fixed our issue.

Co-authored-by: Esteban Torres <me+github@estebantorr.es>